requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
Bottleneck==1.6.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
courlan==1.3.2
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.2
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
elevenlabs==2.27.0
fastapi==0.104.1
fonttools==4.61.1
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-aiplatform==1.115.0
google-cloud-bigquery==3.39.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-speech==2.35.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.33.0
google-cloud-translate==3.23.0
google-crc32c==1.8.0
google-genai==1.4.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.81
langchain-google-vertexai==2.1.2
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==3.0.1
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.15
langsmith==0.5.2
logtail-python==0.3.4
lxml==6.0.2
lxml_html_clean==0.4.3
marshmallow==3.26.2
matplotlib==3.10.8
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numexpr==2.14.1
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.57.0
pluggy==1.6.0
podcastparser==0.6.11
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-asyncio>=0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
qdrant-client==1.15.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
sentry-sdk==2.48.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
psycopg2-binary>=2.9.0
alembic>=1.13.0
starlette==0.27.0
tenacity==9.1.2
tiktoken==0.12.0
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.24.0
uvloop==0.22.1
validators==0.35.0
watchfiles==1.1.1
websockets==14.2
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.12.8
zstandard==0.25.0
